pyproject.toml
[tool.poetry]
name = "lavamusic-py"
version = "0.1.0"
description = "A simple discord music bot using wavelink"
authors = ["appujet <sdipedit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
wavelink = "^3.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 125
skip-string-normalization = true


[tool.isort]
profile = "black"
combine_as_imports = true
combine_star = true
line_length = 125


[tool.pyright]
reportUnnecessaryTypeIgnoreComment = "warning"
pythonVersion = "3.11"
strictParameterNoneValue = false


include = [
    "src/**",
]


exclude = [
    "**/__pycache__",
    "build",
    "dist"
]